Refactor product form logic and update admin layout
Moved product form logic to a custom hook (useProductFormLogic) for better separation of concerns and maintainability. Added AdminModalContext and wrapped AdminLayouts with AdminModalProvider. Updated InventoryManagement table to improve product/variant display. Adjusted recommendation picture and mega menu grid styles. Changed payment route to 'checkout' in RouterSetup.
