+++ src/components/Admin/Form/ProductForm.jsx
@@ -1,199 +1,40 @@
-import React, { useState, useEffect, useContext } from "react";
-import { useForm, Controller, useFieldArray } from "react-hook-form";
-import { useNavigate, useLoaderData } from "react-router-dom";
-import { createProduct, updateProduct, deleteSlider, deleteProductVariant } from "@/api/api";
-import { MoveLeft, Plus, Save, RotateCcw, Trash } from "lucide-react";
-import toast from "react-hot-toast";
+import React from "react";
+import { Controller } from "react-hook-form";
+import { MoveLeft, Plus, Save, Trash } from "lucide-react";
 import ProductImages from "./FormComponents/ProductImages";
 import DebugInfo from "../DebugInfo";
-import { AdminAPIContext } from "@/context/AdminAPIContext";
+import { useProductFormLogic } from "./useProductFormLogic";
 
 const ProductForm = () => {
-    const navigate = useNavigate();
-    const productData = useLoaderData();
-    const isEditing = !!productData;
-
-    const { categories, brands, countries, suppliers, colors, sizes, materials, units, loading, error } = useContext(AdminAPIContext);
-
-    const [isLoading, setIsLoading] = useState(false);
-    const [images, setImages] = useState([]);
-    const [imageErrors, setImageErrors] = useState({});
-    const [deletingVariant, setDeletingVariant] = useState(null); // Trạng thái xóa biến thể
-
     const {
         control,
         handleSubmit,
-        formState: { errors, isSubmitting },
-        reset,
+        errors,
+        isSubmitting,
         watch,
-    } = useForm({
-        defaultValues: isEditing
-            ? {
-                  Id: productData.id,
-                  ProductName: productData.productName || "",
-                  SpecificationDescription: productData.specificationDescription || "",
-                  BrandId: productData.brandId || "",
-                  CategoryId: productData.categoryId || "",
-                  CountriesId: productData.countriesId || "",
-                  SupplierId: productData.supplierId || "",
-                  IsActive: productData.isActive || true,
-                  ProductVariants: productData.productVariants?.map((pv) => ({
-                      Id: pv.id || 0, // Ánh xạ id từ API sang Id trong form
-                      OriginalPrice: pv.originalPrice || 0,
-                      DiscountedPrice: pv.discountedPrice || null,
-                      StockQty: pv.stockQty || 0,
-                      ColorId: pv.colorId || "",
-                      SizeId: pv.sizeId || "",
-                      MaterialId: pv.materialId || "",
-                      UnitId: pv.unitId || "",
-                  })) || [
-                      {
-                          Id: 0,
-                          OriginalPrice: 0,
-                          StockQty: 0,
-                          ColorId: "",
-                          SizeId: "",
-                          MaterialId: "",
-                          UnitId: "",
-                      },
-                  ],
-              }
-            : {
-                  ProductName: "",
-                  SpecificationDescription: "",
-                  BrandId: "",
-                  CategoryId: "",
-                  CountriesId: "",
-                  SupplierId: "",
-                  IsActive: true,
-                  ProductVariants: [
-                      {
-                          Id: 0,
-                          OriginalPrice: 0,
-                          StockQty: 0,
-                          ColorId: "",
-                          SizeId: "",
-                          MaterialId: "",
-                          UnitId: "",
-                      },
-                  ],
-              },
-    });
-
-    const { fields, append, remove } = useFieldArray({
-        control,
-        name: "ProductVariants",
-    });
-
-    useEffect(() => {
-        if (isEditing && productData.sliders) {
-            setImages(
-                productData.sliders.map((slider) => ({
-                    id: slider.id,
-                    file: null,
-                    preview: slider.imageUrl,
-                    alt: slider.title || "",
-                    displayOrder: slider.displayOrder || 0,
-                })),
-            );
-        } else {
-            // Không tạo khung ảnh trống khi tạo mới
-            setImages([]);
-        }
-    }, [isEditing, productData]);
-
-    const validateImages = () => {
-        const newErrors = {};
-        if (images.length === 0 || images.every((img) => !img.file && !img.preview)) {
-            newErrors.images = "Please upload at least one image";
-        }
-        setImageErrors(newErrors);
-        return Object.keys(newErrors).length === 0;
-    };
-
-    // Hàm xử lý xóa biến thể
-    const handleDeleteVariant = async (index) => {
-        const variantId = watch(`ProductVariants[${index}].Id`);
-
-        if (variantId > 0) {
-            // Yêu cầu xác nhận trước khi xóa biến thể đã lưu
-            if (!window.confirm("Are you sure you want to delete this variant?")) {
-                return;
-            }
-
-            // Biến thể đã tồn tại trong backend, gọi API để xóa
-            setDeletingVariant(index); // Đặt trạng thái đang xóa
-            try {
-                await deleteProductVariant(variantId);
-                remove(index); // Xóa khỏi danh sách form sau khi xóa thành công
-                toast.success("Product variant deleted successfully.");
-            } catch (error) {
-                console.error(`Error deleting variant with ID ${variantId}:`, error.message);
-                toast.error(`Error: ${error.message}`);
-            } finally {
-                setDeletingVariant(null); // Reset trạng thái
-            }
-        } else {
-            // Biến thể chưa được lưu, chỉ xóa khỏi form
-            remove(index);
-            toast.success("Product variant removed from form.");
-        }
-    };
-
-    const onSubmit = async (data) => {
-        if (isLoading || !validateImages()) return;
-        setIsLoading(true);
-        try {
-            const productPayload = {
-                ...data,
-                ProductVariants: data.ProductVariants.map((pv) => ({
-                    ...pv,
-                    OriginalPrice: Number(pv.OriginalPrice),
-                    DiscountedPrice: pv.DiscountedPrice ? Number(pv.DiscountedPrice) : null,
-                    StockQty: Number(pv.StockQty),
-                    ColorId: Number(pv.ColorId),
-                    SizeId: Number(pv.SizeId),
-                    MaterialId: Number(pv.MaterialId),
-                    UnitId: Number(pv.UnitId),
-                })),
-            };
-
-            // Hàm định dạng số thứ tự thành chuỗi 3 chữ số (ví dụ: 1 -> "001")
-            const formatSliderIndex = (index) => String(index + 1).padStart(3, "0");
-
-            const sliders = images.map((img, index) => ({
-                imageFile: img.file,
-                title: img.alt,
-                linkUrl: `/products/${isEditing ? productData.id : "new"}`,
-                displayOrder: index,
-            }));
-
-            if (isEditing) {
-                await updateProduct(
-                    productPayload,
-                    sliders.filter((s) => s.imageFile),
-                );
-                toast.success("Product updated successfully!");
-            } else {
-                await createProduct(
-                    productPayload,
-                    sliders.filter((s) => s.imageFile),
-                );
-                toast.success("Product created successfully!");
-            }
-            navigate("/admin/products");
-        } catch (error) {
-            toast.error(`Error: ${error.message}`);
-        } finally {
-            setIsLoading(false);
-        }
-    };
-
-    const resetForm = () => {
-        reset();
-        setImages([]); // Không tạo khung ảnh trống khi reset
-        setImageErrors({});
-    };
+        fields,
+        append,
+        remove,
+        isLoading,
+        images,
+        setImages,
+        imageErrors,
+        setImageErrors,
+        deletingVariant,
+        handleDeleteVariant,
+        onSubmit,
+        isEditing,
+        navigate,
+        categories,
+        brands,
+        countries,
+        suppliers,
+        colors,
+        sizes,
+        materials,
+        units,
+        deleteSlider,
+    } = useProductFormLogic();
 
     return (
         <div className="min-h-screen">
