+++ src/pages/layouts/AdminLayouts.jsx
@@ -3,12 +3,12 @@ import NavbarAdmin from "@/components/Admin/NavbarAdmin";
 import { ThemeProvider } from "@/context/ThemeContext";
 import { cn } from "@/lib/utils";
 import { useMediaQuery } from "@uidotdev/usehooks";
-import { Dessert } from "lucide-react";
 import React, { useEffect, useRef, useState } from "react";
 import { Outlet } from "react-router-dom";
 import { useClickOutside } from "@/hooks/use-click-outside";
 import { Toaster } from "react-hot-toast";
 import { AdminAPIProvider } from "@/context/AdminAPIContext";
+import { AdminModalProvider } from "@/context/AdminModalContext";
 
 const AdminLayouts = ({ children }) => {
     // Kiểm tra breakpoint là màn hình desktop
@@ -29,40 +29,42 @@ const AdminLayouts = ({ children }) => {
     });
 
     return (
-        <ThemeProvider>
-            <AdminAPIProvider>
-                <div className="min-h-screen bg-slate-100 transition-colors dark:bg-slate-950">
-                    {/* Nền đen */}
-                    <div
-                        className={cn(
-                            `pointer-events-none fixed inset-0 -z-10 bg-black opacity-0 transition-opacity`,
-                            !collapsed && "max-md:pointer-events-auto max-md:z-50 max-md:opacity-30",
-                        )}
-                    />
-
-                    {/* 1. Thanh Sidebar */}
-                    <AppSidebar
-                        ref={sidebarRef}
-                        collapsed={collapsed}
-                    />
-
-                    {/* 2. Phần Navbar và Content */}
-                    <div className={cn(`transition-[margin] duration-300`, collapsed ? "md:ml-[70px]" : "md:ml-[240px]")}>
-                        {/* 2.1: Header Navbar */}
-                        <NavbarAdmin
-                            collapsed={collapsed}
-                            setCollapsed={setCollapsed}
+        <AdminModalProvider>
+            <ThemeProvider>
+                <AdminAPIProvider>
+                    <div className="min-h-screen bg-slate-100 transition-colors dark:bg-slate-950">
+                        {/* Nền đen */}
+                        <div
+                            className={cn(
+                                `pointer-events-none fixed inset-0 -z-10 bg-black opacity-0 transition-opacity`,
+                                !collapsed && "max-md:pointer-events-auto max-md:z-50 max-md:opacity-30",
+                            )}
                         />
 
-                        {/* 2.2: Phần nội dung */}
-                        <div className="h-[calc(100vh-60px)] overflow-x-hidden overflow-y-auto p-6">
-                            <Outlet />
-                            <Toaster toastOptions={{ duration: 2000 }} /> {/* 5. Thiết lập thông báo Hot Toast Mặc định */}
+                        {/* 1. Thanh Sidebar */}
+                        <AppSidebar
+                            ref={sidebarRef}
+                            collapsed={collapsed}
+                        />
+
+                        {/* 2. Phần Navbar và Content */}
+                        <div className={cn(`transition-[margin] duration-300`, collapsed ? "md:ml-[70px]" : "md:ml-[240px]")}>
+                            {/* 2.1: Header Navbar */}
+                            <NavbarAdmin
+                                collapsed={collapsed}
+                                setCollapsed={setCollapsed}
+                            />
+
+                            {/* 2.2: Phần nội dung */}
+                            <div className="h-[calc(100vh-60px)] overflow-x-hidden overflow-y-auto p-6">
+                                <Outlet />
+                                <Toaster toastOptions={{ duration: 2000 }} /> {/* 5. Thiết lập thông báo Hot Toast Mặc định */}
+                            </div>
                         </div>
                     </div>
-                </div>
-            </AdminAPIProvider>
-        </ThemeProvider>
+                </AdminAPIProvider>
+            </ThemeProvider>
+        </AdminModalProvider>
     );
 };
